chore: bump go-header and go-fraud (#2619)
This PR bumps go-header and go-fraud which are both breaking - lots of
changes here so please review carefully.

Note that this also contains re-introduction of
VerifyCommitLightTrusting which caused us issues in `mocha` due to rapid
voting power changes. At the moment, there are no test utilities to test
that specific case, so we need to wait to roll it out to `mocha`.

Related: https://github.com/celestiaorg/celestia-app/issues/2382

---------

+++ das/worker.go
@@ -135,7 +135,7 @@ func (w *worker) sample(ctx context.Context, timeout time.Duration, height uint6
 	if w.state.job.jobType == recentJob {
 		err = w.broadcast(ctx, shrexsub.Notification{
 			DataHash: h.DataHash.Bytes(),
-			Height:   uint64(h.Height()),
+			Height:   h.Height(),
 		})
 		if err != nil {
 			log.Warn("failed to broadcast availability message",
